requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
pytest==8.3.3
fastembed==0.7.1
